+++ packages/rum-core/src/domain/assembly.ts
@@ -1,6 +1,5 @@
 import {
   combine,
-  Configuration,
   Context,
   isEmptyObject,
   limitModification,
@@ -32,6 +31,7 @@ import { LifeCycle, LifeCycleEventType } from './lifeCycle'
 import { ParentContexts } from './parentContexts'
 import { RumSessionManager, RumSessionPlan } from './rumSessionManager'
 import { UrlContexts } from './urlContexts'
+import { RumConfiguration } from './configuration'
 
 enum SessionType {
   SYNTHETICS = 'synthetics',
@@ -59,8 +59,7 @@ const OTHER_EVENTS_MODIFIABLE_FIELD_PATHS = [
 type Mutable<T> = { -readonly [P in keyof T]: T[P] }
 
 export function startRumAssembly(
-  applicationId: string,
-  configuration: Configuration,
+  configuration: RumConfiguration,
   lifeCycle: LifeCycle,
   sessionManager: RumSessionManager,
   parentContexts: ParentContexts,
@@ -84,7 +83,10 @@ export function startRumAssembly(
     ({ startTime, rawRumEvent, domainContext, savedCommonContext, customerContext }) => {
       const viewContext = parentContexts.findView(startTime)
       const urlContext = urlContexts.findUrl(startTime)
-      const session = sessionManager.findTrackedSession(startTime)
+      // allow to send events if the session was tracked when they start
+      // except for views which are continuously updated
+      // TODO: stop sending view updates when session is expired
+      const session = sessionManager.findTrackedSession(rawRumEvent.type !== RumEventType.VIEW ? startTime : undefined)
       if (session && viewContext && urlContext) {
         const actionContext = parentContexts.findAction(startTime)
         const commonContext = savedCommonContext || getCommonContext()
@@ -98,7 +100,7 @@ export function startRumAssembly(
             browser_sdk_version: canUseEventBridge() ? buildEnv.sdkVersion : undefined,
           },
           application: {
-            id: applicationId,
+            id: configuration.applicationId,
           },
           date: timeStampNow(),
           service: configuration.service,