+++ packages/rum-core/test/specHelper.ts
@@ -1,8 +1,6 @@
 import {
   assign,
   combine,
-  Configuration,
-  DEFAULT_CONFIGURATION,
   Observable,
   TimeStamp,
   noop,
@@ -27,13 +25,14 @@ import {
   SYNTHETICS_TEST_ID_COOKIE_NAME,
 } from '../src/domain/syntheticsContext'
 import type { CiTestWindow } from '../src/domain/ciTestContext'
+import { RumConfiguration, validateAndBuildRumConfiguration } from '../src/domain/configuration'
 import { validateFormat } from './formatValidation'
 import { createRumSessionManagerMock } from './mockRumSessionManager'
 
 export interface TestSetupBuilder {
   withFakeLocation: (initialUrl: string) => TestSetupBuilder
   withSessionManager: (sessionManager: RumSessionManager) => TestSetupBuilder
-  withConfiguration: (overrides: Partial<Configuration>) => TestSetupBuilder
+  withConfiguration: (overrides: Partial<RumConfiguration>) => TestSetupBuilder
   withParentContexts: (stub: Partial<ParentContexts>) => TestSetupBuilder
   withForegroundContexts: (stub: Partial<ForegroundContexts>) => TestSetupBuilder
   withFakeClock: () => TestSetupBuilder
@@ -48,7 +47,7 @@ export interface BuildContext {
   lifeCycle: LifeCycle
   domMutationObservable: Observable<void>
   locationChangeObservable: Observable<LocationChange>
-  configuration: Readonly<Configuration>
+  configuration: Readonly<RumConfiguration>
   sessionManager: RumSessionManager
   location: Location
   applicationId: string
@@ -93,11 +92,11 @@ export function setup(): TestSetupBuilder {
     selectInForegroundPeriodsFor: () => undefined,
     stop: noop,
   }
-  const configuration: Partial<Configuration> = {
-    ...DEFAULT_CONFIGURATION,
+  const FAKE_APP_ID = 'appId'
+  const configuration: RumConfiguration = {
+    ...validateAndBuildRumConfiguration({ clientToken: 'xxx', applicationId: FAKE_APP_ID })!,
     ...SPEC_ENDPOINTS,
   }
-  const FAKE_APP_ID = 'appId'
 
   // ensure that events generated before build are collected
   const rawRumEventsCollected = lifeCycle.subscribe(LifeCycleEventType.RAW_RUM_EVENT_COLLECTED, (data) => {
@@ -123,7 +122,7 @@ export function setup(): TestSetupBuilder {
       sessionManager = sessionManagerStub
       return setupBuilder
     },
-    withConfiguration(overrides: Partial<Configuration>) {
+    withConfiguration(overrides: Partial<RumConfiguration>) {
       assign(configuration, overrides)
       return setupBuilder
     },
@@ -154,7 +153,7 @@ export function setup(): TestSetupBuilder {
           foregroundContexts,
           sessionManager,
           applicationId: FAKE_APP_ID,
-          configuration: configuration as Configuration,
+          configuration,
           location: fakeLocation as Location,
         })
         if (result && result.stop) {